Tweak current ruby version for TruffleRuby
