+++ spec/core/core_spec.rb
@@ -6,7 +6,7 @@ describe "refined patches" do
   it "activates only unsupported patches" do
     next skip unless RubyNext::Utils.refine_modules?
 
-    ruby_version = Gem::Version.new(RUBY_VERSION)
+    ruby_version = Gem::Version.new(::RubyNext.current_ruby_version)
 
     activated_patches = RubyNext::Core.patches.refined.values.flatten.uniq
